+++ language-server/src/tests/action.rs
@@ -3847,6 +3847,24 @@ pub fn main(x) -> option.Option(Int) {
 }
 
 #[test]
+fn test_qualified_to_unqualified_only_triggers_within_qualified_value() {
+    let src = r#"
+import option
+
+pub fn main(x) -> option.Option(Int) {
+    option.Some(1)
+}
+// end
+"#;
+    assert_no_code_actions!(
+        "Unqualify option.Option",
+        TestProject::for_source(src)
+            .add_hex_module("option", "pub type Option(v) { Some(v) None }"),
+        find_position_of("import").select_until(find_position_of("// end")),
+    );
+}
+
+#[test]
 fn test_qualified_to_unqualified_import_nested_type_outer() {
     let src = r#"
 import option
@@ -4503,7 +4521,25 @@ pub fn main() {
     assert_code_action!(
         "Qualify map as list.map",
         TestProject::for_source(src).add_hex_module("list", "pub fn map(list, f) { todo }"),
-        find_position_of("map(").select_until(find_position_of("[1, 2, 3]")),
+        find_position_of("map").nth_occurrence(2).to_selection(),
+    );
+}
+
+#[test]
+fn test_unqualified_to_qualified_only_triggers_when_within_an_expression() {
+    let src = r#"
+import list.{map}
+
+pub fn main() {
+    let identity = map([1, 2, 3], fn(x) { x })
+    let double = map([1, 2, 3], fn(x) { x * 2 })
+}
+// end
+"#;
+    assert_no_code_actions!(
+        "Qualify map as list.map",
+        TestProject::for_source(src).add_hex_module("list", "pub fn map(list, f) { todo }"),
+        find_position_of("import").select_until(find_position_of("// end")),
     );
 }
 
@@ -4523,7 +4559,7 @@ pub fn circle_circumference(radius: Float) -> Float {
     assert_code_action!(
         "Qualify pi as mymath.pi",
         TestProject::for_source(src).add_hex_module("mymath", "pub const pi = 3.14159"),
-        find_position_of("pi *.").select_until(find_position_of(" radius")),
+        find_position_of("pi").nth_occurrence(2).to_selection(),
     );
 }
 
@@ -4540,7 +4576,10 @@ pub fn create_user(name: String) -> User {
         "Qualify User as user.User",
         TestProject::for_source(src)
             .add_hex_module("user", "pub type User { User(name: String, id: Int) }"),
-        find_position_of("User(").select_until(find_position_of("name: name")),
+        find_position_of("User")
+            .nth_occurrence(4)
+            .under_char('s')
+            .to_selection(),
     );
 }
 
@@ -4557,7 +4596,7 @@ import user.{type User, User}
         "Qualify User as user.User",
         TestProject::for_source(src)
             .add_hex_module("user", "pub type User { User(name: String, id: Int) }"),
-        find_position_of("User(").select_until(find_position_of("name: name")),
+        find_position_of("User").nth_occurrence(2).to_selection(),
     );
 }
 
@@ -4580,7 +4619,7 @@ pub fn user_list(users: List(User)) -> List(String) {
         "Qualify User as user.User",
         TestProject::for_source(src)
             .add_hex_module("user", "pub type User { User(name: String, id: Int) }"),
-        find_position_of("User(").select_until(find_position_of("name: name")),
+        find_position_of("User").nth_occurrence(2).to_selection(),
     );
 }
 
@@ -4601,7 +4640,7 @@ pub fn process_list(items: List(Int)) -> List(Int) {
             "list",
             "pub fn map(list: List(a), with fun: fn(a) -> b) -> List(b) { todo }"
         ),
-        find_position_of("|> map").select_until(find_position_of("(fn(x)")),
+        find_position_of("map").nth_occurrence(2).to_selection(),
     );
 }
 
@@ -4621,7 +4660,7 @@ pub fn process_result(res: Result(Int, String)) -> Int {
         "Qualify Ok as result.Ok",
         TestProject::for_source(src)
             .add_hex_module("result", "pub type Result(a, e) { Ok(a) Error(e) }"),
-        find_position_of("Ok(").select_until(find_position_of("value)")),
+        find_position_of("Ok").nth_occurrence(2).to_selection(),
     );
 }
 
@@ -4643,7 +4682,7 @@ pub fn maybe_increment(x: Option(Int)) -> Option(Int) {
             .add_hex_module("option", "pub type Option(a) { Some(a) None }"),
         find_position_of("Opt")
             .nth_occurrence(2)
-            .select_until(find_position_of("ion(")),
+            .select_until(find_position_of("ion").nth_occurrence(3)),
     );
 }
 
@@ -4668,7 +4707,9 @@ pub fn process_names(names: List(List(Int))) -> List(Int) {
 pub fn flatten(lists: List(List(a))) -> List(a) { todo }"
             )
             .add_hex_module("operation", "pub fn double(s: Int) -> Int { todo }"),
-        find_position_of("(dou").select_until(find_position_of("ble)")),
+        find_position_of("dou")
+            .nth_occurrence(2)
+            .select_until(find_position_of("ble").nth_occurrence(2)),
     );
 }
 
@@ -4687,7 +4728,9 @@ pub fn double_list(items: List(Int)) -> List(Int) {
             "list",
             "pub fn map(list: List(a), with fun: fn(a) -> b) -> List(b) { todo }"
         ),
-        find_position_of("transform(").select_until(find_position_of("items,")),
+        find_position_of("transform")
+            .nth_occurrence(2)
+            .to_selection(),
     );
 }
 
@@ -4706,7 +4749,9 @@ pub fn double_list(items: List(Int)) -> List(Int) {
             "list",
             "pub fn map(list: List(a), with fun: fn(a) -> b) -> List(b) { todo }"
         ),
-        find_position_of("transform(").select_until(find_position_of("items,")),
+        find_position_of("transform")
+            .nth_occurrence(2)
+            .to_selection(),
     );
 }
 
@@ -4726,7 +4771,9 @@ pub fn double_list(items: List(Int)) -> List(Int) {
             "list",
             "pub fn map(list: List(a), with fun: fn(a) -> b) -> List(b) { todo }"
         ),
-        find_position_of("transform(").select_until(find_position_of("items,")),
+        find_position_of("transform")
+            .nth_occurrence(2)
+            .to_selection(),
     );
 }
 
@@ -4748,7 +4795,7 @@ pub fn maybe_increment(x: Option(Int)) -> Option(Int) {
             .add_hex_module("option", "pub type Option(a) { Some(a) None }"),
         find_position_of("Opt")
             .nth_occurrence(2)
-            .select_until(find_position_of("ion(")),
+            .select_until(find_position_of("ion").nth_occurrence(3)),
     );
 }
 
@@ -4770,7 +4817,7 @@ pub fn maybe_increment(x: Maybe(Int)) -> Maybe(Int) {
             .add_hex_module("option", "pub type Option(a) { Some(a) None }"),
         find_position_of("May")
             .nth_occurrence(2)
-            .select_until(find_position_of("be(")),
+            .select_until(find_position_of("be").nth_occurrence(3)),
     );
 }
 
@@ -4792,7 +4839,7 @@ pub fn maybe_increment(x: Option(Int)) -> Option(Int) {
             .add_hex_module("option", "pub type Option(a) { Some(a) None }"),
         find_position_of("Opt")
             .nth_occurrence(2)
-            .select_until(find_position_of("ion(")),
+            .select_until(find_position_of("ion").nth_occurrence(3)),
     );
 }
 
@@ -4810,7 +4857,7 @@ pub fn main() {
         "Qualify Some as option.Some",
         TestProject::for_source(src)
             .add_hex_module("option", "pub type Option(v) { Some(v) None }"),
-        find_position_of("Some(").select_until(find_position_of("1)")),
+        find_position_of("Some(1)").select_until(find_position_of("Some(1)").under_char('e')),
     );
 }
 #[test]
@@ -4835,7 +4882,7 @@ pub fn main() {
         "Qualify Some as option.Some",
         TestProject::for_source(src)
             .add_hex_module("option", "pub type Option(v) { Some(v) None }"),
-        find_position_of("Some(").select_until(find_position_of("1)")),
+        find_position_of("Some(1)").select_until(find_position_of("Some(1)").under_char('e')),
     );
 }
 
@@ -5707,6 +5754,17 @@ fn inexhaustive_let_alias_to_case() {
   let 10 as ten = 10
 }",
         find_position_of("let").select_until(find_position_of("=")),
+    );
+}
+
+#[test]
+fn cursor_must_be_within_let_assignment_to_trigger_action() {
+    assert_no_code_actions!(
+        CONVERT_TO_CASE,
+        "pub fn main() {
+  let 10 as ten = 10
+}",
+        find_position_of("pub").select_until(find_position_of("}")),
     );
 }
 
@@ -7668,6 +7726,24 @@ pub type Person {
 }
 
 #[test]
+fn generate_dynamic_decoder_only_works_if_withing_a_type() {
+    assert_no_code_actions!(
+        GENERATE_DYNAMIC_DECODER,
+        "
+pub type Person {
+  Person(name: String, age: Int, height: Float, is_cool: Bool, brain: BitArray)
+}
+
+pub fn main() {
+  // unrelated
+  todo
+}
+",
+        find_position_of("pub type").select_until(find_position_of("todo"))
+    );
+}
+
+#[test]
 fn generate_dynamic_decoder_complex_types() {
     let src = "
 import gleam/option
@@ -8706,6 +8782,30 @@ pub type Person {
             "pub type Json"
         ),
         find_position_of("type").to_selection()
+    );
+}
+
+#[test]
+fn generate_json_only_triggers_within_a_type() {
+    let src = "
+pub type Person {
+  Person(name: String, age: Int, height: Float, is_cool: Bool)
+}
+
+pub fn main() {
+  // unrelated
+  todo
+}
+";
+
+    assert_no_code_actions!(
+        GENERATE_TO_JSON_FUNCTION,
+        TestProject::for_source(src).add_package_module(
+            "gleam_json",
+            "gleam/json",
+            "pub type Json"
+        ),
+        find_position_of("type").select_until(find_position_of("todo"))
     );
 }
 
@@ -10100,6 +10200,19 @@ fn wibble() {
 }
 "#,
         find_position_of("case").select_until(find_position_of("True {"))
+    );
+}
+
+#[test]
+fn add_missing_patterns_needs_to_be_within_the_inexhaustive_case_expression() {
+    assert_no_code_actions!(
+        ADD_MISSING_PATTERNS,
+        r#"
+fn wibble() {
+  case True {}
+}
+"#,
+        find_position_of("fn").select_until(find_position_of("}"))
     );
 }
 
@@ -11793,6 +11906,77 @@ pub fn main() {
 }
 
 #[test]
+fn no_extract_function_selecting_multiple_case_branches() {
+    assert_no_code_actions!(
+        EXTRACT_FUNCTION,
+        r#"
+const pi = 3.14
+
+pub fn main() {
+  let value = 3.15
+
+  let string = case value {
+    0.0 -> "Zero"
+    1.0 -> "One"
+    _ -> "Something else"
+  }
+
+  echo string
+}
+"#,
+        find_position_of("0.0").select_until(find_position_of("else"))
+    );
+}
+
+#[test]
+fn no_extract_function_selecting_case_branch_pattern() {
+    assert_no_code_actions!(
+        EXTRACT_FUNCTION,
+        r#"
+const pi = 3.14
+
+pub fn main() {
+  let value = 3.15
+
+  let string = case value {
+    0.0 -> "Zero"
+    1.0 -> "One"
+    _ -> "Something else"
+  }
+
+  echo string
+}
+"#,
+        find_position_of("0.0").select_until(find_position_of("Zero").under_last_char())
+    );
+}
+
+#[test]
+fn no_extract_function_selecting_case_branch_guard() {
+    assert_no_code_actions!(
+        EXTRACT_FUNCTION,
+        r#"
+const pi = 3.14
+
+pub fn main() {
+  let value = 3.15
+
+  let string = case value {
+    0.0 if True -> "Zero"
+    1.0 -> "One"
+    _ -> "Something else"
+  }
+
+  echo string
+}
+"#,
+        find_position_of("True")
+            .under_char('u')
+            .select_until(find_position_of("Zero").under_last_char())
+    );
+}
+
+#[test]
 fn extract_use_inside_function() {
     assert_code_action!(
         EXTRACT_FUNCTION,
@@ -12768,6 +12952,24 @@ type Wibble {
 }
 
 #[test]
+fn add_missing_type_parameter_can_only_trigger_if_within_type() {
+    assert_no_code_actions!(
+        ADD_MISSING_TYPE_PARAMETER,
+        r#"
+type Wibble {
+  Wibble(field: t)
+}
+
+pub fn main() {
+    // unrelated
+    todo
+}
+"#,
+        find_position_of("type").select_until(find_position_of("todo"))
+    );
+}
+
+#[test]
 fn add_missing_type_parameter_to_exising_parameter() {
     assert_code_action!(
         ADD_MISSING_TYPE_PARAMETER,
@@ -13682,6 +13884,24 @@ fn op(i: Int) -> Int {
 }
 ",
         find_position_of("fn(int)").to_selection()
+    );
+}
+
+#[test]
+fn unwrap_anonymous_function_can_only_trigger_if_cursor_is_within_the_function() {
+    assert_no_code_actions!(
+        UNWRAP_ANONYMOUS_FUNCTION,
+        "import gleam/list
+
+pub fn main() {
+  list.map([1, 2, 3], fn(int) { op(int) })
+}
+
+fn op(i: Int) -> Int {
+  todo
+}
+",
+        find_position_of("fn(int)").select_until(find_position_of("todo"))
     );
 }